Some refactor. Empty unit test file for join plans
